requirements.txt
# Tech Stack - Quản Lý Dịch Vụ Cho Thuê Kho Lưu Trữ Hàng Hóa
# Cập nhật: 23/04/2026
# Yêu cầu: Python 3.10+

# Core - GUI Framework
PyQt6>=6.4.0

# Database & ORM
SQLAlchemy>=2.0.0

# PDF Generation - Xuất báo cáo, hóa đơn
reportlab>=3.6.0

# Utilities
python-dateutil>=2.8.0

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0          # Code formatter
# Excel Export
pandas>=2.0.0
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0  # For coverage reports
